Refactored maps domain.
