+++ packages/engine/src/core/index.ts
@@ -2,4 +2,4 @@ export * from './terraforming-costs';
 export * from './game-context';
 export * from './game';
 
-export * from './maps/basic-map-model';
+export * from '../maps/basic-map-model';
